Write filters to query log from default editor.
