+++ src/plugins/data/public/index.ts
@@ -420,6 +420,7 @@ export {
   syncQueryStateWithUrl,
   QueryState,
   getDefaultQuery,
+  getQueryLog,
   FilterManager,
   SavedQuery,
   SavedQueryService,
