+++ src/plugins/data/public/query/lib/get_query_log.ts
@@ -22,7 +22,6 @@ import { IStorageWrapper } from 'src/plugins/kibana_utils/public';
 import { PersistedLog } from '../persisted_log';
 import { UI_SETTINGS } from '../../../common';
 
-/** @internal */
 export function getQueryLog(
   uiSettings: IUiSettingsClient,
   storage: IStorageWrapper,
